requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
astor=0.8.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=pypi_0
certifi=2021.10.8=py37h03978a9_1
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
debugpy=1.5.1=py37hf2a7229_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=2.10.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
ipykernel=6.9.2=py37h90c5f73_0
ipython=7.32.0=py37h03978a9_0
jedi=0.18.1=py37h03978a9_0
joblib=1.1.0=pypi_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py37h03978a9_0
kafka-python=2.0.2=pypi_0
keras=2.2.5=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
libsodium=1.0.18=h8d14728_1
markdown=3.3.6=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
nltk=3.7=pypi_0
numpy=1.21.5=pypi_0
openssl=1.1.1n=h8ffe710_0
opt-einsum=3.3.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=21.2.4=py37haa95532_0
prompt-toolkit=3.0.27=pyha770c72_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=py37hcc03f2d_0
pygments=2.11.2=pyhd8ed1ab_0
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pywin32=303=py37hcc03f2d_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py37hcce574b_1
regex=2022.3.15=pypi_0
requests=2.27.1=pypi_0
scipy=1.7.3=pypi_0
setuptools=58.0.4=py37haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.0=h2bbff1b_0
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
tornado=6.1=py37hcc03f2d_2
tqdm=4.63.1=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py37haa95532_2
wrapt=1.14.0=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.7.0=pypi_0
